backend/requirements.txt
Django==2.2.19
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv==0.20.0
sqlparse==0.3.1
django-cors-headers==3.13.0
drf-extra-fields==3.4.0
Pillow==9.2.0
django-filter==21.1